Extracted stateful logic from svg renderer.
Added LineWalker as an example of rules and rendering combined, this has the stateful logic e.g. maintaining transformation that was previous combined with the svg rendering.

This separation seems a lot better, probably a bit more work to do on a reasonable standard interface for these systems + their "rendering" of output.